requirements.txt
# Core ML framework
torch>=2.2.0
torchvision>=0.17.0

# HuggingFace ecosystem
transformers>=4.49.0
datasets>=3.0.0
accelerate>=1.0.0
huggingface_hub>=0.27.0
tokenizers>=0.21.0

# Tokenizer dependencies (required for multilingual models like tiny-aya-base)
sentencepiece>=0.2.0
protobuf>=4.25.0

# Image processing (required for SigLIP 2 vision encoder)
Pillow>=10.0.0
requests>=2.32.0

# Numerical computing
numpy>=1.26.0

# Utilities
tqdm>=4.66.0
einops>=0.8.0

# Parameter-efficient fine-tuning (Phase 2+)
peft>=0.13.0

# Evaluation
lm-eval>=0.4.0

# Quantization (optional, for tight GPU memory)
bitsandbytes>=0.43.0

# Dev tools
ruff>=0.8.0
pytest>=8.0.0